HSEARCH-3049 Move DocumentCommitStrategy and DocumentRefreshStrategy to the org.hibernate.search.backend.work.execution package
For consistency with the previous commit. Keeping this in a separate
commit since it's an API change.
