+++ backend/lucene/src/test/java/org/hibernate/search/backend/lucene/orchestration/impl/LuceneWriteWorkProcessorTest.java
@@ -17,8 +17,8 @@ import java.util.concurrent.CompletableFuture;
 import org.hibernate.search.backend.lucene.lowlevel.writer.impl.IndexWriterDelegator;
 import org.hibernate.search.backend.lucene.work.impl.LuceneWriteWork;
 import org.hibernate.search.backend.lucene.work.impl.LuceneWriteWorkExecutionContext;
-import org.hibernate.search.engine.backend.index.DocumentCommitStrategy;
-import org.hibernate.search.engine.backend.index.DocumentRefreshStrategy;
+import org.hibernate.search.engine.backend.work.execution.DocumentCommitStrategy;
+import org.hibernate.search.engine.backend.work.execution.DocumentRefreshStrategy;
 import org.hibernate.search.engine.common.spi.ContextualErrorHandler;
 import org.hibernate.search.engine.common.spi.ErrorHandler;
 import org.hibernate.search.engine.reporting.spi.EventContexts;
