+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/session/AutomaticIndexingSynchronizationStrategy.java
@@ -8,8 +8,8 @@ package org.hibernate.search.mapper.orm.session;
 
 import java.util.concurrent.CompletableFuture;
 
-import org.hibernate.search.engine.backend.index.DocumentCommitStrategy;
-import org.hibernate.search.engine.backend.index.DocumentRefreshStrategy;
+import org.hibernate.search.engine.backend.work.execution.DocumentCommitStrategy;
+import org.hibernate.search.engine.backend.work.execution.DocumentRefreshStrategy;
 import org.hibernate.search.util.common.impl.Futures;
 
 /**
